feat: added a loading state using SSE
